testing bert

+++ experiments/task72/test_bert.py
@@ -40,7 +40,7 @@ def create_submition_file(outfolder, mymodel, loader, device):
         csv_writer = csv.writer(output, delimiter=',',
                                      quotechar='"', quoting=csv.QUOTE_MINIMAL)
         for id,out in zip(ids_list, outs_list):
-            csv_writer.writerow([id, int(out)])
+            csv_writer.writerow([id, float(out)])
 
 
 DEVICE = 'cuda' if torch.cuda.is_available() else 'cpu'
